+++ frontend/src/features/editors/boxEditor/functionBoxes/FunctionContent.tsx
@@ -1,4 +1,4 @@
-import styles from "../BoxEditorStyles.module.css";
+import styles from "../../Editor.module.css";
 import { ID } from "../../../shared/types";
 import IdSelector from "../../idEditor/IdEditor";
