+++ frontend/src/features/editors/classEditor/ClassEditor.module.css
