+++ frontend/src/features/editors/idEditor/IdDisplay.tsx
@@ -1,4 +1,4 @@
-import styles from "../boxEditor/BoxEditorStyles.module.css";
+import styles from "../Editor.module.css";
 import { ID } from "../../shared/types";
 import IdSelector from "./IdEditor";
